LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/KeyframeEffect-getKeyframes.tentative-expected.txt


FAIL KeyframeEffect.getKeyframes() returns expected frames for a simple transition assert_equals: value for 'easing' on ComputedKeyframe #0 expected "linear" but got "ease"
FAIL KeyframeEffect.getKeyframes() returns frames unaffected by a non-default easing function assert_equals: value for 'easing' on ComputedKeyframe #0 expected "linear" but got "steps(2)"
FAIL KeyframeEffect.getKeyframes() returns expected frames for a transition with a CSS variable endpoint assert_equals: value for 'easing' on ComputedKeyframe #0 expected "linear" but got "ease"
